+++ src/GafferUIModule/ViewportGadgetBinding.cpp
@@ -128,7 +128,7 @@ struct UnarySlotCaller
 		}
 		catch( const error_already_set &e )
 		{
-			PyErr_PrintEx( 0 ); // clears the error status
+			IECorePython::ExceptionAlgo::translatePythonException();
 		}
 	}
 };